requirements.txt
aiogram==3.4.1
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
aiosqlite==0.19.0
pydantic==2.6.1